Move SetupSuccessSubtleNotification to the ads module.
